Creates a separate package for RPS functions.
