+++ internal/handler/games/rps.go
@@ -1,22 +1,14 @@
 package games
 
 import (
-	"encoding/json"
 	"fmt"
-	"github.com/pyrousnet/pyrous-gobot/internal/cache"
 	"github.com/pyrousnet/pyrous-gobot/internal/comms"
-	"reflect"
+	"github.com/pyrousnet/pyrous-gobot/internal/handler/games/rps"
 	"strings"
 
 	"github.com/google/uuid"
 	"github.com/pyrousnet/pyrous-gobot/internal/users"
 )
-
-type RPS struct {
-	RpsPlaying string `json:"rps-playing"`
-	Rps        string `json:"rps"`
-	Name       string `json:"name"`
-}
 
 func (bg BotGame) Rps(event BotGame) error {
 	response := comms.Response{
@@ -31,15 +23,21 @@ func (bg BotGame) Rps(event BotGame) error {
 	} else {
 		foundChannel = event.ReplyChannel
 	}
-	playerUser, _, err := users.GetUser(strings.TrimLeft(event.sender, "@"), event.cache)
+	playerUser, _, err := users.GetUser(strings.TrimLeft(event.sender, "@"), event.Cache)
+	response.UserId = playerUser.Id
 	if err != nil {
 		return err
 	}
-	player, perr := getPlayer(playerUser, foundChannel.Id, event.cache)
-	opponent, oErr := findApponent(event, player, foundChannel.Id)
-	if perr != nil || !playing(player) {
-		if oErr == nil && playing(opponent) {
-			channelId, ok, _ := event.cache.Get(opponent.RpsPlaying)
+	player, perr := rps.GetPlayer(playerUser, foundChannel.Id, event.Cache)
+	rpsEvent := rps.RpsBotGame{
+		ReplyChannel:    event.ReplyChannel,
+		ResponseChannel: event.ResponseChannel,
+		Cache:           event.Cache,
+	}
+	opponent, oErr := rps.FindApponent(rpsEvent, player, foundChannel.Id)
+	if perr != nil || !rps.Playing(player) {
+		if oErr == nil && rps.Playing(opponent) {
+			channelId, ok, _ := event.Cache.Get(opponent.RpsPlaying)
 			if ok && event.ReplyChannel != nil && channelId == foundChannel.Id {
 				player.RpsPlaying = opponent.RpsPlaying
 				response.Type = "dm"
@@ -52,7 +50,7 @@ func (bg BotGame) Rps(event BotGame) error {
 				Type:   "dm",
 				UserId: playerUser.Id,
 			}
-			event.cache.Put(id.String(), event.ReplyChannel.Id)
+			event.Cache.Put(id.String(), event.ReplyChannel.Id)
 			response.Message = fmt.Sprintf("/echo %s is looking for an opponent in RPS.", event.sender)
 			dmResponse.Message = fmt.Sprintf("Would you like to throw Rock, Paper or Scissors (Usage: $rps %s rock)", event.ReplyChannel.Name)
 			event.ResponseChannel <- dmResponse
@@ -78,9 +76,9 @@ func (bg BotGame) Rps(event BotGame) error {
 	}
 
 	if oErr == nil && opponent.Name != "" {
-		winners, hasWinner := getWinner(player, opponent)
+		winners, hasWinner := rps.GetWinner(player, opponent)
 		if hasWinner {
-			channelId, ok, _ := event.cache.Get(player.RpsPlaying)
+			channelId, ok, _ := event.Cache.Get(player.RpsPlaying)
 			response.Type = "command"
 			if ok {
 				response.ReplyChannelId = channelId.(string)
@@ -91,155 +89,19 @@ func (bg BotGame) Rps(event BotGame) error {
 				}
 			}
 
-			deleteGame(player.RpsPlaying, event.cache)
-			deleteRps(player, foundChannel.Id, event.cache)
-			deleteRps(opponent, foundChannel.Id, event.cache)
+			rps.DeleteGame(player.RpsPlaying, event.Cache)
+			rps.DeleteRps(player, foundChannel.Id, event.Cache)
+			rps.DeleteRps(opponent, foundChannel.Id, event.Cache)
 			event.ResponseChannel <- response
 			return err
 		}
 
-		updateRps(opponent, foundChannel.Id, event.cache)
+		rps.UpdateRps(opponent, foundChannel.Id, event.Cache)
 	}
 
-	updateRps(player, foundChannel.Id, event.cache)
+	rps.UpdateRps(player, foundChannel.Id, event.Cache)
 
 	event.ResponseChannel <- response
 
 	return err
-}
-
-func playing(player RPS) bool {
-	return player.RpsPlaying != ""
-}
-
-func sameGame(player RPS, opponent RPS) bool {
-	return player.RpsPlaying == "" || player.RpsPlaying == opponent.RpsPlaying
-}
-
-func differentUser(player RPS, opponent RPS) bool {
-	return player.Name != opponent.Name
-}
-
-func findApponent(event BotGame, forPlayer RPS, chanId string) (RPS, error) {
-	us, ok, err := users.GetUsers(event.cache)
-	rpsUs, ok, gPerr := getPlayers(us, chanId, event.cache)
-	var opponent RPS
-	var found = false
-
-	if us == nil {
-		return RPS{}, fmt.Errorf("no opponent")
-	}
-
-	if ok {
-		for _, u := range rpsUs {
-			if playing(u) && sameGame(forPlayer, u) && differentUser(forPlayer, u) {
-				opponent = u
-				found = true
-			}
-		}
-	} else {
-		return RPS{}, gPerr
-	}
-
-	if !found {
-		err = fmt.Errorf("no opponent")
-	}
-	return opponent, err
-}
-
-func getWinner(player RPS, opponent RPS) ([]RPS, bool) {
-	var hasWinner bool = false
-	var winners []RPS
-
-	if player.Rps == "" {
-		return []RPS{}, false
-	}
-
-	if opponent.Rps == "" {
-		return []RPS{}, false
-	}
-
-	if player.Rps == opponent.Rps {
-		winners = append(winners, player)
-		winners = append(winners, opponent)
-		hasWinner = true
-	} else if strings.ToLower(player.Rps) == "rock" {
-		if strings.ToLower(opponent.Rps) == "paper" {
-			winners = append(winners, opponent)
-			hasWinner = true
-		} else if strings.ToLower(opponent.Rps) == "scissors" {
-			winners = append(winners, player)
-			hasWinner = true
-		}
-	} else if strings.ToLower(player.Rps) == "paper" {
-		if strings.ToLower(opponent.Rps) == "scissors" {
-			winners = append(winners, opponent)
-			hasWinner = true
-		} else if strings.ToLower(opponent.Rps) == "rock" {
-			winners = append(winners, player)
-			hasWinner = true
-		}
-	} else if strings.ToLower(player.Rps) == "scissors" {
-		if strings.ToLower(opponent.Rps) == "rock" {
-			winners = append(winners, opponent)
-			hasWinner = true
-		} else if strings.ToLower(opponent.Rps) == "paper" {
-			winners = append(winners, player)
-			hasWinner = true
-		}
-	}
-
-	return winners, hasWinner
-}
-
-func getPlayer(player users.User, chanId string, c cache.Cache) (RPS, error) {
-	key := fmt.Sprintf("%s-%s-%s", "rps", player.Name, chanId)
-	var rps RPS
-	r, ok, _ := c.Get(key)
-	if ok {
-		if reflect.TypeOf(r).String() != "[]uint8" {
-			json.Unmarshal([]byte(r.(string)), &rps)
-		} else {
-			json.Unmarshal(r.([]byte), &rps)
-		}
-		return rps, nil
-	}
-	return RPS{Name: player.Name}, fmt.Errorf("not found")
-}
-
-func getPlayers(pUsers []users.User, chanId string, c cache.Cache) ([]RPS, bool, error) {
-	var rpsUsers []RPS
-	for _, u := range pUsers {
-		key := fmt.Sprintf("%s-%s-%s", "rps", u.Name, chanId)
-		r, ok, _ := c.Get(key)
-		var rps RPS
-		if ok {
-			if reflect.TypeOf(r).String() != "[]uint8" {
-				json.Unmarshal([]byte(r.(string)), &rps)
-			} else {
-				json.Unmarshal(r.([]byte), &rps)
-			}
-
-			rpsUsers = append(rpsUsers, rps)
-		}
-	}
-	if len(rpsUsers) == 0 {
-		return []RPS{}, false, nil
-	}
-	return rpsUsers, true, nil
-}
-
-func updateRps(playerRps RPS, chanId string, c cache.Cache) (RPS, error) {
-	key := fmt.Sprintf("%s-%s-%s", "rps", playerRps.Name, chanId)
-	p, _ := json.Marshal(playerRps)
-	c.Put(key, p)
-	return playerRps, nil
-}
-
-func deleteRps(playerRps RPS, chanId string, c cache.Cache) {
-	key := fmt.Sprintf("%s-%s-%s", "rps", playerRps.Name, chanId)
-	c.Clean(key)
-}
-func deleteGame(uuid string, c cache.Cache) {
-	c.Clean(uuid)
 }
